+++ job/urls.py
@@ -37,6 +37,7 @@ urlpatterns = [
     re_path(r'^job/inputs/?$', views.inputs),
     re_path(r'^job/outputs/?$', views.outputs),
     re_path(r'^chp_defaults/?$', views.chp_defaults),
+    re_path(r'^absorption_chiller_defaults/?$', views.absorption_chiller_defaults),
     re_path(r'^emissions_profile/?$', views.emissions_profile),
     re_path(r'^easiur_costs/?$', views.easiur_costs),
     re_path(r'^simulated_load/?$', views.simulated_load),